Quantum Tensor Networks

Qrochet.Quantum — Type

Quantum

Tensor Network with a notion of "causality". This leads to the notion of sites and directionality (input/output).

Notes

Indices are referenced by Sites.

Missing docstring.

Missing docstring for Tenet.TensorNetwork(::Quantum). Check Documenter's build log for details.

Base.adjoint — Method

adjoint(q::Quantum)

Returns the adjoint of a Quantum Tensor Network; i.e. the conjugate Tensor Network with the inputs and outputs swapped.

Qrochet.sites — Function

sites(q::Quantum)

Returns the sites of a Quantum Tensor Network.

Qrochet.nsites — Function

nsites(q::Quantum)

Returns the number of sites of a Quantum Tensor Network.

Queries

Missing docstring.

Missing docstring for Tenet.inds(::Quantum; kwargs...). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Tenet.tensors(::Quantum; kwargs...). Check Documenter's build log for details.

Connecting Quantum Tensor Networks

Qrochet.inputs — Function

inputs(q::Quantum)

Returns the input sites of a Quantum Tensor Network.

Qrochet.outputs — Function

outputs(q::Quantum)

Returns the output sites of a Quantum Tensor Network.

Missing docstring.

Missing docstring for lanes. Check Documenter's build log for details.

Qrochet.ninputs — Function

ninputs(q::Quantum)

Returns the number of input sites of a Quantum Tensor Network.

Qrochet.noutputs — Function

noutputs(q::Quantum)

Returns the number of output sites of a Quantum Tensor Network.

Missing docstring.

Missing docstring for nlanes. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Socket. Check Documenter's build log for details.

Qrochet.socket — Method

socket(q::Quantum)

Returns the socket of a Quantum Tensor Network; i.e. whether it is a Scalar, State or Operator.

Qrochet.Scalar — Type

Scalar <: Socket

Socket representing a scalar; i.e. a Tensor Network with no open sites.

Qrochet.State — Type

State <: Socket

Socket representing a state; i.e. a Tensor Network with only input sites (or only output sites if dual = true).

Qrochet.Operator — Type

Operator <: Socket

Socket representing an operator; i.e. a Tensor Network with both input and output sites.

Base.merge — Method

merge(a::Quantum, b::Quantum...)

Merges multiple Quantum Tensor Networks into a single one by connecting input/output sites.